requirements.txt
py-cord
python-dotenv
mysql-connector-python
pytz
asyncio